Change flask to simple http fileserver
The simple file server allows serving files over http requests with
less code than flask. The html content is provided over requests
with any path following url '/'.

The directory is not available with python 3.5.3 so the current working
directory will be used to resolve the working directory of the http
server. The folders are added to gitignore to avoid seeing them as
untracked files.